+++ lib/generators/doorkeeper/templates/initializer.rb
@@ -25,6 +25,9 @@ Doorkeeper.configure do
   #   Admin.find_by_id(session[:admin_id]) || redirect_to(routes.new_admin_session_url)
   # end
 
+  # Authorization Code expiration time (default 10 minutes).
+  # access_token_expires_in 10.minutes
+
   # Access token expiration time (default 2 hours).
   # If you want to disable expiration, set this to nil.
   # access_token_expires_in 2.hours